rename custom_tool to client_tool and other changes
