+++ examples/agents/inflation.py
@@ -16,10 +16,10 @@ from termcolor import colored
 
 
 def run_main(host: str, port: int, disable_safety: bool = False):
-    if "BRAVE_SEARCH_API_KEY" not in os.environ:
+    if "TAVILY_SEARCH_API_KEY" not in os.environ:
         print(
             colored(
-                "Warning: BRAVE_SEARCH_API_KEY is not set; will not use Search tool.",
+                "Warning: TAVILY_SEARCH_API_KEY is not set; will not use websearch tool.",
                 "yellow",
             )
         )
@@ -50,12 +50,7 @@ def run_main(host: str, port: int, disable_safety: bool = False):
         },
         instructions="You are a helpful assistant",
         toolgroups=(
-            (
-                ["builtin::websearch"]
-                if os.getenv("BRAVE_SEARCH_API_KEY")
-                or os.getenv("TAVILY_SEARCH_API_KEY")
-                else []
-            )
+            (["builtin::websearch"] if os.getenv("TAVILY_SEARCH_API_KEY") else [])
             + ["builtin::code_interpreter"]
         ),
         tool_choice="required",
